change IProviderAdapter to IProviderContext
