+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/IProvider.java
@@ -1,4 +1,23 @@
 package org.eclipse.jdt.ls.debug.adapter;
 
-public interface IProvider extends  IConfigurable {
+import java.util.HashMap;
+import java.util.Map;
+
+public interface IProvider {
+    /**
+     * Initialize this provider.
+     * @param options the options
+     */
+    default void initialize(Map<String, Object> options) {
+
+    }
+
+    /**
+     * Get the default options for this provider.
+     *
+     * @return The default options.
+     */
+    default Map<String, Object> getDefaultOptions() {
+        return new HashMap<>();
+    }
 }
