--- org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/IProviderAdapter.java
@@ -1,8 +1,0 @@
-package org.eclipse.jdt.ls.debug.adapter;
-
-public interface IProviderAdapter {
-    
-    <T extends IProvider> T getProvider(Class<T> clazz);
-
-    void registerProvider(Class<? extends IProvider> clazz, IProvider provider);
-}
