+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/IProviderContext.java
@@ -1,19 +1,12 @@
-/*******************************************************************************
-* Copyright (c) 2017 Microsoft Corporation and others.
-* All rights reserved. This program and the accompanying materials
-* are made available under the terms of the Eclipse Public License v1.0
-* which accompanies this distribution, and is available at
-* http://www.eclipse.org/legal/epl-v10.html
-*
-* Contributors:
-*     Microsoft Corporation - initial API and implementation
-*******************************************************************************/
-
 package org.eclipse.jdt.ls.debug.adapter;
 
 public interface IProviderContext {
+    
+    <T extends IProvider> T getProvider(Class<T> clazz);
+
+    void registerProvider(Class<? extends IProvider> clazz, IProvider provider);
 
     ISourceLookUpProvider getSourceLookUpProvider();
-    
+
     IVirtualMachineManagerProvider getVirtualMachineManagerProvider();
 }
